fix: Added SameSite=None config (2)
