+++ src/main/java/com/classvar/common/config/TomcatConfig.java
@@ -4,6 +4,9 @@ import org.apache.catalina.Context;
 import org.apache.catalina.connector.Connector;
 import org.apache.tomcat.util.descriptor.web.SecurityCollection;
 import org.apache.tomcat.util.descriptor.web.SecurityConstraint;
+import org.apache.tomcat.util.http.Rfc6265CookieProcessor;
+import org.apache.tomcat.util.http.SameSiteCookies;
+import org.springframework.boot.web.embedded.tomcat.TomcatContextCustomizer;
 import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
 import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
 import org.springframework.context.annotation.Bean;
@@ -11,6 +14,18 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class TomcatConfig {
+
+    // SameSite=None을 추가하는 필터
+    // SameSite=None 옵션은 HTTPS에서만 작동한다.
+    // Secure 옵션도 지정돼야 하는데, Spring Boot에서 Session 설정이 기본으로 httpOnly, Secure = true로 되어있는 듯하다.
+    @Bean
+    public TomcatContextCustomizer sameSiteCookiesConfig() {
+        return context -> {
+            final Rfc6265CookieProcessor cookieProcessor = new Rfc6265CookieProcessor();
+            cookieProcessor.setSameSiteCookies(SameSiteCookies.NONE.getValue());
+            context.setCookieProcessor(cookieProcessor);
+        };
+    }
 
     @Bean
     public ServletWebServerFactory servletContainer() {
